feat: migrate to latest api and remove inference sdk version
