+++ README.md
@@ -21,7 +21,7 @@
 
 </div>
 
-This sample shows how to deploy a secure [Azure AI Services](https://learn.microsoft.com/azure/ai-services/what-are-ai-services) infrastructure with DeepSeek-R1 model, along with reusable components to build a web UI with authentication. It provides a starting point for building secure AI chat applications, using [RBAC](https://learn.microsoft.com/azure/role-based-access-control/built-in-roles) permissions and Azure AI Inference SDK with [keyless (Entra) authentication](https://learn.microsoft.com/entra/identity/managed-identities-azure-resources/overview). The backend resources are secured within an [Azure Virtual Network](https://learn.microsoft.com/azure/virtual-network/virtual-networks-overview), and the frontend is hosted on [Azure Static Web Apps](https://learn.microsoft.com/azure/static-web-apps/overview).
+This sample shows how to deploy a secure [Azure AI Services](https://learn.microsoft.com/azure/ai-services/what-are-ai-services) infrastructure with DeepSeek-R1 model, along with reusable components to build a web UI with authentication. It provides a starting point for building secure AI chat applications, using [RBAC](https://learn.microsoft.com/azure/role-based-access-control/built-in-roles) permissions and OpenAI SDK with [keyless (Entra) authentication](https://learn.microsoft.com/entra/identity/managed-identities-azure-resources/overview). The backend resources are secured within an [Azure Virtual Network](https://learn.microsoft.com/azure/virtual-network/virtual-networks-overview), and the frontend is hosted on [Azure Static Web Apps](https://learn.microsoft.com/azure/static-web-apps/overview).
 
 > [!NOTE]
 > The DeepSeek-R1 model focus is on complex reasoning tasks, and it is not designed for general conversation. It is best suited for tasks that require a deep understanding of the context and a complex reasoning process to provide an answer.
@@ -41,7 +41,7 @@ This application is made from multiple components:
 
 - Example web app integrations of the web components, hosted on [Azure Static Web Apps](https://learn.microsoft.com/azure/static-web-apps/overview). There are example using [static HTML](./packages/webapp-html/), [React](./packages/webapp-react/), [Angular](./packages/webapp-angular/), [Vue](./packages/webapp-vue/) and [Svelte](./packages/webapp-svelte/).
 
-- A serverless API built with [Azure Functions](https://learn.microsoft.com/azure/azure-functions/functions-overview?pivots=programming-language-javascript) and using either [Azure AI Inference SDK](https://www.npmjs.com/package/@azure-rest/ai-inference) or [OpenAI SDK](https://www.npmjs.com/package/openai) to generate responses to the user chat queries. The code is located in the `packages/api` folder.
+- A serverless API built with [Azure Functions](https://learn.microsoft.com/azure/azure-functions/functions-overview?pivots=programming-language-javascript) and using [OpenAI SDK](https://www.npmjs.com/package/openai) to generate responses to the user chat queries. The code is located in the `packages/api` folder.
 
 We use the [HTTP protocol for AI chat apps](https://aka.ms/chatprotocol) to communicate between the web app and the API.
 
@@ -165,64 +165,28 @@ Open the URL `http://localhost:4280` in your browser, use the authentication emu
 
 ## Key learnings
 
-To use Azure AI Services in your application, you can use the [Azure AI Inference SDK](https://www.npmjs.com/package/@azure-rest/ai-inference) or the [OpenAI SDK](https://www.npmjs.com/package/openai).
+To use Azure AI Services in your application, you can use the [OpenAI SDK](https://www.npmjs.com/package/openai).
 
-Both of these SDKs supports [Managed Identity](https://learn.microsoft.com/azure/active-directory/managed-identities-azure-resources/overview) authentication, which allows you to authenticate without using any secrets. This is a great way to secure your application and avoid hardcoding secrets in your code.
-
-### Using Azure AI Inference SDK
-
-```ts
-import { DefaultAzureCredential } from '@azure/identity';
-import ModelClient, { isUnexpected } from '@azure-rest/ai-inference';
-
-const credentialScopes = {
-  credentials: { scopes: ['https://cognitiveservices.azure.com/.default'] }
-};
-const endpoint = 'https://<YOUR_INSTANCE_NAME>.services.ai.azure.com/models';
-
-// Use the current user identity to authenticate.
-// No secrets needed, it uses `az login` or `azd auth login` locally,
-// and managed identity when deployed on Azure.
-const credentials = new DefaultAzureCredential();
-const client = ModelClient(endpoint, credentials, credentialScopes);
-
-const response = await client.path('/chat/completions').post({
-  body: {
-    messages: [{ role: 'user', content: 'Say hello' }],
-    model: 'DeepSeek-R1',
-  },
-});
-
-console.log(response.body.choices[0].message.content);
-```
+This SDK supports [Managed Identity](https://learn.microsoft.com/azure/active-directory/managed-identities-azure-resources/overview) authentication, which allows you to authenticate without using any secrets. This is a great way to secure your application and avoid hardcoding secrets in your code.
 
 ### Using OpenAI SDK
 
-> [!IMPORTANT]
-> Currently, only the v5 beta version of the OpenAI SDK supports managed identity authentication with Azure AI Service. You can install it using the following command: `npm install openai@beta`.
-
 ```ts
 import { DefaultAzureCredential, getBearerTokenProvider } from '@azure/identity';
-import OpenAI from 'openai';
+import { AzureOpenAI } from 'openai';
 
 const scope = 'https://cognitiveservices.azure.com/.default';
-const endpoint = 'https://<YOUR_INSTANCE_NAME>.services.ai.azure.com/models';
+const endpoint = 'https://<YOUR_INSTANCE_NAME>.services.ai.azure.com';
 
 // Use the current user identity to authenticate.
 // No secrets needed, it uses `az login` or `azd auth login` locally,
 // and managed identity when deployed on Azure.
 const credentials = new DefaultAzureCredential();
 const azureADTokenProvider = getBearerTokenProvider(credentials, scope);
-const client = new OpenAI({
-  baseURL: endpoint,
-  fetch: async (url, init = {}) => {
-    const token = await azureADTokenProvider();
-    const headers = new Headers(init.headers);
-    headers.set('Authorization', `Bearer ${token}`);
-    return fetch(url, { ...init, headers });
-  },
-  defaultQuery: { 'api-version': '2024-05-01-preview' },
-  apiKey: '_not_used_',
+const client = new AzureOpenAI({
+  endpoint,
+  azureADTokenProvider,
+  apiVersion: '2025-04-01-preview',
 });
 
 const response = await client.chat.completions.create({
@@ -233,14 +197,15 @@ const response = await client.chat.completions.create({
 console.log(response.choices[0].message.content);
 ```
 
-### Parsing DeepSeek-R1 responses
+### DeepSeek-R1 responses and reasoning
 
-The DeepSeek-R1 model generates responses in a specific format. The response contains a `<think>` tag that the model uses for its reasoning, and the final answer is provided after the `</think>` tag. You can use the following code to extract the reasoning and the final answer from the response:
+The DeepSeek-R1 model generates responses in a specific format. The response contains a `reasoning_content` property that the model uses for its reasoning, and the final answer is provided after within the `content` property. You can use the following code to extract the reasoning and the final answer from the response:
 
 ```ts
-const [thoughts, content] = message.content
-  .match(/<think>(.*?)<\/think>(.*)/s)?
-  .slice(1) ?? [];
+const content = response.choices[0].message.content;
+// Note that reasoning_content property is not yet supported in the OpenAI SDK types,
+// so we need to cast the message to access it.
+const reasoningContent = (response.choices[0].message as any).reasoning_content;
 ```
 
 ## Resources
