+++ packages/api/package.json
@@ -14,13 +14,12 @@
   "author": "Microsoft",
   "license": "MIT",
   "dependencies": {
-    "@azure-rest/ai-inference": "^1.0.0-beta.6",
     "@azure/core-sse": "^2.1.3",
     "@azure/functions": "^4.0.0",
     "@azure/identity": "^4.2.0",
     "@microsoft/ai-chat-protocol": "^1.0.0-beta.20240814.1",
     "dotenv": "^16.4.5",
-    "openai": "^5.0.0-beta.0"
+    "openai": "^5.1.0"
   },
   "devDependencies": {
     "@types/node": "^20.x",
