+++ packages/api/src/functions/chat-post.ts
@@ -1,17 +1,17 @@
 import process from 'node:process';
 import { Readable } from 'node:stream';
-import { DefaultAzureCredential } from '@azure/identity';
 import { type HttpRequest, type InvocationContext, type HttpResponseInit, app } from '@azure/functions';
+import { DefaultAzureCredential, getBearerTokenProvider } from '@azure/identity';
+import { AzureOpenAI } from 'openai';
+import { type ChatCompletionChunk } from 'openai/resources/chat';
 import {
   type AIChatCompletionRequest,
   type AIChatCompletionDelta,
   type AIChatCompletion,
 } from '@microsoft/ai-chat-protocol';
-import ModelClient, { isUnexpected } from '@azure-rest/ai-inference';
-import { createSseStream, NodeJSReadableStream } from '@azure/core-sse';
 import 'dotenv/config';
 
-const credentialScopes = { credentials: { scopes: ['https://cognitiveservices.azure.com/.default'] } };
+const credentialScope = 'https://cognitiveservices.azure.com/.default';
 const systemPrompt = `Assistant helps the user with cooking questions. Be brief in your answers. Answer only plain text, DO NOT use Markdown.
 
 After your answer, ALWAYS generate 3 very brief follow-up questions that the user would likely ask next, based on the context.
@@ -55,31 +55,22 @@ export async function postChat(
     // No secrets needed, it uses `az login` or `azd auth login` locally,
     // and managed identity when deployed on Azure.
     const credentials = new DefaultAzureCredential();
-    const client = ModelClient(endpoint, credentials, credentialScopes);
+    const azureADTokenProvider = getBearerTokenProvider(credentials, credentialScope);
+    const client = new AzureOpenAI({
+      endpoint,
+      azureADTokenProvider,
+      apiVersion: '2025-04-01-preview',
+    });
 
     if (stream) {
-      const responseStream = await client
-        .path('/chat/completions')
-        .post({
-          body: {
-            messages: [{ role: 'system', content: systemPrompt }, ...messages],
-            temperature: 0.7,
-            model,
-            stream: true,
-          },
-        })
-        .asNodeStream();
+      const responseStream = await client.chat.completions.create({
+        messages: [{ role: 'system', content: systemPrompt }, ...messages],
+        temperature: 0.7,
+        model,
+        stream: true,
+      });
 
-      const stream = responseStream.body;
-      if (!stream) {
-        throw new Error('The response stream is undefined');
-      }
-
-      if (responseStream.status !== '200') {
-        throw new Error('Failed to get chat completions');
-      }
-
-      const jsonStream = Readable.from(createJsonStream(stream));
+      const jsonStream = Readable.from(createJsonStream(responseStream));
       return {
         headers: {
           'Content-Type': 'application/x-ndjson',
@@ -89,23 +80,21 @@ export async function postChat(
       };
     }
 
-    const response = await client.path('/chat/completions').post({
-      body: {
-        messages: [{ role: 'system', content: systemPrompt }, ...messages],
-        temperature: 0.7,
-        model,
-      },
+    const response = await client.chat.completions.create({
+      messages: [{ role: 'system', content: systemPrompt }, ...messages],
+      temperature: 0.7,
+      model,
     });
-
-    if (isUnexpected(response)) {
-      throw new Error(response.body.error.message);
-    }
 
     return {
       jsonBody: {
         message: {
-          content: response.body.choices[0].message.content,
+          content: response.choices[0].message.content,
           role: 'assistant',
+          context: {
+            // reasoning_content property is not yet part of the type definition
+            reasoning: (response.choices[0].message as any).reasoning_content ?? '',
+          },
         },
       } as AIChatCompletion,
     };
@@ -120,21 +109,19 @@ export async function postChat(
   }
 }
 
-// Transform the SSE stream into a JSON stream
-async function* createJsonStream(stream: NodeJSReadableStream) {
-  const events = createSseStream(stream);
-
-  for await (const event of events) {
-    if (event.data === '[DONE]') return;
-    if (!event.data) continue;
-
-    const chunk = JSON.parse(event.data);
-    if (!chunk.choices[0]?.delta.content) continue;
+// Transform the response chunks into a JSON stream
+async function* createJsonStream(chunks: AsyncIterable<ChatCompletionChunk>) {
+  for await (const chunk of chunks) {
+    if (!chunk || !chunk.choices[0]) continue;
 
     const responseChunk: AIChatCompletionDelta = {
       delta: {
-        content: chunk.choices[0].delta.content as string,
+        content: chunk.choices[0]?.delta.content ?? '',
         role: 'assistant',
+        context: {
+          // reasoning_content property is not yet part of the type definition
+          reasoning: (chunk.choices[0]?.delta as any).reasoning_content ?? '',
+        },
       },
     };
